reactor move uiState files to store/editor/ directory
