+++ src/store/editorStorage.test.ts
@@ -18,7 +18,7 @@
 import { type EditorState } from "@/pageEditor/store/editor/pageEditorTypes";
 import { initialState } from "@/pageEditor/store/editor/editorSlice";
 import { type UUID } from "@/types/stringTypes";
-import { type BrickConfigurationUIState } from "@/pageEditor/uiState/uiStateTypes";
+import { type BrickConfigurationUIState } from "@/pageEditor/store/editor/uiStateTypes";
 import { getPipelineMap } from "@/pageEditor/tabs/editTab/editHelpers";
 import {
   getEditorState,
