+++ src/tsconfig.strictNullChecks.json
@@ -1208,8 +1208,8 @@
     "./pageEditor/tabs/trigger/TriggerConfiguration.tsx",
     "./pageEditor/toolbar/PermissionsToolbar.tsx",
     "./pageEditor/toolbar/ReloadToolbar.tsx",
-    "./pageEditor/uiState/uiState.ts",
-    "./pageEditor/uiState/uiStateTypes.ts",
+    "./pageEditor/store/editor/uiState.ts",
+    "./pageEditor/store/editor/uiStateTypes.ts",
     "./pageEditor/utils.ts",
     "./pageScript/elementInfo.ts",
     "./pageScript/frameworks/adapters.ts",
